+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-Change Log for currtime
+Change Log for timeprint
 ================================================================================
 
 ## v2.0.0  In Progress
-- Add help information on new (C++11) codes
+- Add help information on new (C++11) codes.
 - Moved custom elapse codes to use new underscore modifier.
   + Old %D is replaced with %_d,
     %D is now standard "Short MM/DD/YY date, equivalent to %m/%d/%y".
@@ -10,6 +10,7 @@ Change Log for currtime
     %R is now standard "24-hour HH:MM time, equivalent to %H:%M".
   + Old %s is replaced with %_s,
     %s is now undefined.
+- Renamed from currtime to timeprint.
 
 ## v1.0.0  (original 2013-03-09, released 2018-02-15)
   - Original version snap as of 2013-03-09.